fix(content): break resize storage adapter cycle
