+++ src/content/shared/panels/sessionPanelResizeTypes.ts
@@ -1,0 +1,8 @@
+export interface SessionPanelResizeStorage {
+  load(): Promise<Record<string, unknown>>;
+  save(items: Record<string, unknown>): void;
+}
+
+export interface SessionPanelResizeOptions {
+  storage: SessionPanelResizeStorage;
+}
